+++ resources/views/home.blade.php
@@ -16,9 +16,9 @@
 
                     <ul>
                         <li>Nombre de catégories: {{ App\Category::count() }}</li>
-                        <li>Nombre de places: {{ App\Place::count() }}</li>
+                        <li>Nombre de d'entreprises: {{ App\Place::count() }}</li>
                         @if(App\Place::where('is_approved', false)->get()->isNotEmpty())
-                        <li>En attente d'approvation: {{ App\Place::where('is_approved', false)->count() }}. <a href="{{ route("places.moderation") }}">Effectuer la modération d'entreprises par ici.</a></li>
+                        <li>En attente d'approvation: {{ App\Place::where('is_approved', false)->count() }}.</li>
                         @endif
                     </ul>
                 </div>